Fix extra trailing line break with `|-` block literals
The Emitter.write_literal() function will write a line break after a
block literal value if the block literal didn't end with a line break.
The Scanner.scan_block_scalar() function was keeping track of all newlines
that followed block literals, but when `|-` is used it this was one too many.

The code is changed slightly to avoid a confusing `if chomping is not False`
statement.